Remove 760m/s boundary on μ formula (G/Gmax param) in HH model (#35)
* Update helper_hh_calibration.py

We remove the 760m/s Vs limit which was resulting in occasional instabilities for smooth Vs profiles that cross the 760m/s boundary. (correct this time)

* Minor comment edit

Removed the initials and date on the explanatory comment in the code

* Adjusting test corresponding to hh calibration

* Fixing hh calibration test

* Modified test file for param from profile

* Minor formatting edit

* Update test_class_hh_calibration.py benchmarks

Modified the benchmark in case_2_with_both_Vs_profile_and_GGmax_curves()

* Minor formatting edits

* Commas formatting

* Update test for deprecation warning

Addressing DeprecationWarning: Passing None into shape arguments as an alias for () is deprecated.

* Formatting fix

* Python version and changelog edits

* Workflow edit

* Reverting test change

* More workflow changes

* Reverting some changes + pytest adjustment

* Fix testenv

* Changelog and readme edits

* Reverting some changes

* Formatting and test_helper_generic edit